Move secret hooks to a separate file.
