+++ src/juju/api-hooks/index.ts
@@ -1,0 +1,10 @@
+export {
+  useListSecrets,
+  useGetSecretContent,
+  useCreateSecrets,
+  useUpdateSecrets,
+  useRemoveSecrets,
+  useGrantSecret,
+  useRevokeSecret,
+  type Label as SecretsLabel,
+} from "./secrets";
