+++ src/pages/EntityDetails/Model/Secrets/Secrets.tsx
@@ -6,7 +6,7 @@ import { useParams } from "react-router-dom";
 import type { EntityDetailsRoute } from "components/Routes/Routes";
 import useCanManageSecrets from "hooks/useCanManageSecrets";
 import { useQueryParams } from "hooks/useQueryParams";
-import { useListSecrets } from "juju/apiHooks";
+import { useListSecrets } from "juju/api-hooks";
 import { actions as jujuActions } from "store/juju";
 import {
   getModelByUUID,
