+++ src/panels/ConfigPanel/SecretsPicker/SecretsPicker.tsx
@@ -21,7 +21,7 @@ import type { EntityDetailsRoute } from "components/Routes/Routes";
 import SecretForm from "components/secrets/SecretForm";
 import SecretLabel from "components/secrets/SecretLabel";
 import useCanManageSecrets from "hooks/useCanManageSecrets";
-import { useListSecrets } from "juju/apiHooks";
+import { useListSecrets } from "juju/api-hooks";
 import { actions as jujuActions } from "store/juju";
 import {
   getModelByUUID,
