+++ src/panels/GrantSecretPanel/GrantSecretPanel.tsx
@@ -9,7 +9,11 @@ import FormikField from "components/FormikField";
 import Panel from "components/Panel";
 import type { EntityDetailsRoute } from "components/Routes/Routes";
 import SecretLabel from "components/secrets/SecretLabel";
-import { useListSecrets, useGrantSecret, useRevokeSecret } from "juju/apiHooks";
+import {
+  useListSecrets,
+  useGrantSecret,
+  useRevokeSecret,
+} from "juju/api-hooks";
 import PanelInlineErrors from "panels/PanelInlineErrors";
 import { usePanelQueryParams } from "panels/hooks";
 import {
